+++ src/definations/kawaiicon-material/materialDark.ts
@@ -1,0 +1,18 @@
+import fileExtensions from "./fileExtensions";
+import fileNames from "./fileNames";
+import folderNames from "./folderNames";
+import folderNamesExpanded from "./folderNamesExpanded";
+import languageIds from "./languageIds";
+
+export default {
+  file: "file",
+  folder: "folder",
+  folderExpanded: "folder-open",
+  rootFolder: "folder-root",
+  rootFolderExpanded: "folder-root-open",
+  fileExtensions,
+  folderNames,
+  folderNamesExpanded,
+  fileNames,
+  languageIds,
+};