+++ src/definations/kawaiicon-official/kawaiiconDark.ts
@@ -1,0 +1,14 @@
+import fileExtensions from "./fileExtensions";
+import fileNames from "./fileNames";
+import languageIds from "./languageIds";
+
+export default {
+  file: "_file",
+  folder: "_folder",
+  folderExpanded: "_folder_open",
+  rootFolder: "_root_folder",
+  rootFolderExpanded: "_root_folder_open",
+  fileExtensions,
+  fileNames,
+  languageIds,
+};