+++ src/definations/kawaiicon-vscodeicons/vscodeDark.ts
@@ -1,0 +1,18 @@
+import fileExtensions from "./fileExtensions";
+import fileNames from "./fileNames";
+import folderNames from "./folderNames";
+import folderNamesExpanded from "./folderNamesExpanded";
+import languageIds from "./languageIds";
+
+export default {
+  file: "default_file",
+  folder: "default_folder",
+  folderExpanded: "default_folder_opened",
+  rootFolder: "default_root_folder",
+  rootFolderExpanded: "default_root_folder_opened",
+  folderNames,
+  folderNamesExpanded,
+  fileExtensions,
+  fileNames,
+  languageIds,
+};