+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog for Kawaiicon
 
+## 2.0.0 - 2024-08-21
+
+- Added `Material Icon Theme` Icon sets
+- Added `Vscode Icons` Icon sets
+- Other minor changes & bug fixes
+
 ## 1.0.0 - 2024-08-21
 
 - Initial release